refactoring the utils.py into three utils subfiles
